requirements.txt
flask
gunicorn
requests
beautifulsoup4
pyttsx3
Pillow